Improved Tokenizer Logic
The tokenizer is now futureproofed, it was using basic single character checking logic to determine `Tokens` which was no good if you wanted tokens that were more than one character and you had to hack ways in to have multi character tokens. This new tokenizer uses heuristic approach, characters are appended one by one whilst each time this happens it checks whether any of these partially or fully match, if there are partial matches it continues to evaluate until there are no partial or full matches left, at which point it knows that the last full match was the last valid token, so adds it to the list, resets the builder and starts appending again, waiting for the next match.